refactor: restructure climate indicators by flattening modules into the new earthkit.climate.indicators namespace
