--- src/earthkit/climate/atmos/__init__.py
@@ -1,8 +1,0 @@
-# (C) Copyright 2025 - ECMWF and individual contributors.
-
-"""Atmos indicators."""
-
-from earthkit.climate.atmos.precipitation import *  # noqa
-from earthkit.climate.atmos.synoptic import *  # noqa
-from earthkit.climate.atmos.temperature import *  # noqa
-from earthkit.climate.atmos.wind import *  # noqa
