+++ src/earthkit/climate/indicators/__init__.py
@@ -1,0 +1,12 @@
+# (C) Copyright 2025 - ECMWF and individual contributors.
+
+"""Climate indicators."""
+
+from earthkit.climate.indicators.hydrology import *  # noqa
+from earthkit.climate.indicators.land import *  # noqa
+from earthkit.climate.indicators.precipitation import *  # noqa
+from earthkit.climate.indicators.sea_ice import *  # noqa
+from earthkit.climate.indicators.snow import *  # noqa
+from earthkit.climate.indicators.synoptic import *  # noqa
+from earthkit.climate.indicators.temperature import *  # noqa
+from earthkit.climate.indicators.wind import *  # noqa
