+++ src/earthkit/climate/indicators/precipitation.py
