+++ src/earthkit/climate/indicators/sea_ice.py
