+++ src/earthkit/climate/indicators/snow.py
