+++ src/earthkit/climate/indicators/synoptic.py
