+++ src/earthkit/climate/indicators/wind.py
