--- src/earthkit/climate/land/__init__.py
@@ -1,7 +1,0 @@
-# (C) Copyright 2025 - ECMWF and individual contributors.
-
-"""Land indicators."""
-
-from earthkit.climate.land.hydrology import *  # noqa
-from earthkit.climate.land.land import *  # noqa
-from earthkit.climate.land.snow import *  # noqa
