--- src/earthkit/climate/ocean/__init__.py
@@ -1,5 +1,0 @@
-# (C) Copyright 2025 - ECMWF and individual contributors.
-
-"""Sea ice indicators."""
-
-from earthkit.climate.ocean.sea_ice import *  # noqa
