+++ tests/unit/atmos/test_precipitation.py
@@ -14,7 +14,7 @@ import pytest
 import xarray
 from pytest_mock import MockerFixture
 
-from earthkit.climate.atmos import precipitation
+from earthkit.climate.indicators import precipitation
 
 INDICATORS = [
     (precipitation.antecedent_precipitation_index, "antecedent_precipitation_index", {"val": "test"}),
