+++ tests/unit/atmos/test_synoptic.py
@@ -14,7 +14,7 @@ import pytest
 import xarray
 from pytest_mock import MockerFixture
 
-from earthkit.climate.atmos import synoptic
+from earthkit.climate.indicators import synoptic
 
 INDICATORS = [
     (synoptic.jetstream_metric_woollings, "jetstream_metric_woollings"),
